pyproject.toml
[project]
name = "generador-contrasena"
version = "0.1.0"
description = "Generador de contraseñas seguras (Django + JS cliente)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.15.2",
    "django-environ>=0.12.0",
    "django-ratelimit>=4.1.0",
    "django-csp>=4.0",
    "whitenoise>=6.8.2",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
    "django-stubs[compatible-mypy]>=5.1.1",
    "djangorestframework-stubs[compatible-mypy]>=3.15.2",
    "bandit[toml]>=1.8.0",
    "pip-audit>=2.7.3",
    "pre-commit>=4.0.1",
]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = [".venv", "staticfiles", "migrations"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # bugbear
    "C4",   # comprehensions
    "UP",   # pyupgrade
    "S",    # bandit (subset)
    "DJ",   # django
    "SIM",  # simplify
    "RUF",  # ruff-specific
]
ignore = [
    "S101",  # assert in tests
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["S105", "S106", "S311"]
"settings/**" = ["S105"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
exclude = ["\\.venv", "migrations", "staticfiles"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["core.tests.*", "core.migrations.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["environ", "environ.*"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
testpaths = ["core/tests"]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:django.*",
]

[tool.coverage.run]
source = ["core"]
branch = true
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/apps.py",
    "*/admin.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true

[tool.bandit]
exclude_dirs = [".venv", "core/tests", "staticfiles"]
# B101: asserts (handled); B104: 0.0.0.0 bind (dev only); B105: hardcoded test SECRET_KEY.
skips = ["B101", "B104", "B105"]

[tool.djlint]
profile = "django"
indent = 4
max_line_length = 120
